backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.8
sqlalchemy==2.0.35
alembic==1.13.3
python-dotenv==1.0.1
httpx==0.27.2
pydantic-settings==2.5.2
groq==0.11.0
elevenlabs==1.9.0
